+++ packages/expo-sqlite/android/src/main/java/expo/modules/sqlite/SQLiteModule.kt
@@ -45,6 +45,28 @@ class SQLiteModule : Module() {
       return@AsyncFunction pluginResultsToPrimitiveData(results)
     }
 
+    AsyncFunction("execRawQuery") { dbName: String, queries: List<Query>, readOnly: Boolean ->
+      val db = getDatabase(dbName)
+      val results = queries.map { sqlQuery ->
+        val sql = sqlQuery.sql
+        val bindArgs = convertParamsToStringArray(sqlQuery.args)
+        try {
+          if (isSelect(sql)) {
+            doSelectInBackgroundAndPossiblyThrow(sql, bindArgs, db)
+          } else { // update/insert/delete
+            if (readOnly) {
+              SQLitePluginResult(EMPTY_ROWS, EMPTY_COLUMNS, 0, 0, ReadOnlyException())
+            } else {
+              doRawUpdate(sql, bindArgs, db)
+            }
+          }
+        } catch (e: Throwable) {
+          SQLitePluginResult(EMPTY_ROWS, EMPTY_COLUMNS, 0, 0, e)
+        }
+      }
+      return@AsyncFunction pluginResultsToPrimitiveData(results)
+    }
+
     AsyncFunction("close") { dbName: String ->
       DATABASES
         .remove(dbName)
@@ -102,6 +124,78 @@ class SQLiteModule : Module() {
         EMPTY_RESULT
       }
     }
+  }
+
+  private fun doRawUpdate(
+    sql: String,
+    bindArgs: Array<String?>,
+    db: SQLiteDatabase
+  ): SQLitePluginResult {
+    return db.rawQuery(sql, bindArgs).use { cursor ->
+      val numRows = cursor.count
+      if (numRows == 0) {
+        return EMPTY_RESULT
+      }
+
+      val numColumns = cursor.columnCount
+      val columnNames = cursor.columnNames
+      val rows: Array<Array<Any?>> = Array(numRows) { arrayOfNulls(numColumns) }
+      var i = 0
+      while (cursor.moveToNext()) {
+        val row = rows[i]
+        for (j in 0 until numColumns) {
+          row[j] = getValueFromCursor(cursor, j, cursor.getType(j))
+        }
+        rows[i] = row
+        i++
+      }
+
+      if (isInsert(sql)) {
+        val rowsAffected = getRowsAffected(db).let {
+          it.first.close()
+          it.second
+        }
+        val insertId = getInsertId(db).let {
+          it.first.close()
+          it.second
+        }
+        return SQLitePluginResult(rows, columnNames, rowsAffected, insertId, null)
+      } else if (isDelete(sql) || isUpdate(sql)) {
+        val rowsAffected = getRowsAffected(db).let {
+          it.first.close()
+          it.second
+        }
+        SQLitePluginResult(rows, columnNames, rowsAffected, 0, null)
+      } else {
+        EMPTY_RESULT
+      }
+    }
+  }
+
+  private fun getRowsAffected(
+    db: SQLiteDatabase,
+  ): Pair<Cursor, Int> {
+    val cursor = db.rawQuery("SELECT changes() AS numRowsAffected", null)
+    val rowsAffected = if (cursor.moveToFirst()) {
+      val index = cursor.getColumnIndex("numRowsAffected")
+      cursor.getInt(index)
+    } else {
+      -1
+    }
+    return Pair(cursor, rowsAffected)
+  }
+
+  private fun getInsertId(
+    db: SQLiteDatabase,
+  ): Pair<Cursor, Long> {
+    val cursor = db.rawQuery("SELECT last_insert_rowid() AS insertId", null)
+    val insertId = if (cursor.moveToFirst()) {
+      val index = cursor.getColumnIndex("insertId")
+      cursor.getLong(index)
+    } else {
+      -1
+    }
+    return Pair(cursor, insertId)
   }
 
   // do a select operation